# switched to Cholesky matrix decomposition for performance issues