Timed matrix decomposition algorithm: 

1. Cholesky: Time taken: 13.821453291s 
2. ColPivQR: Time taken: 72.457952125s
3. FullPivLU: Time taken: 68.823773959s
4. LU: Time taken: 23.604132917s
5. QR: Time taken: 29.650277583s